+++ awc/tests/utils.rs
@@ -4,8 +4,9 @@
 use std::io::{Read as _, Write as _};
 
 pub mod gzip {
-    use super::*;
     use flate2::{read::GzDecoder, write::GzEncoder, Compression};
+
+    use super::*;
 
     pub fn encode(bytes: impl AsRef<[u8]>) -> Vec<u8> {
         let mut encoder = GzEncoder::new(Vec::new(), Compression::fast());
@@ -22,8 +23,9 @@ pub mod gzip {
 }
 
 pub mod deflate {
-    use super::*;
     use flate2::{read::ZlibDecoder, write::ZlibEncoder, Compression};
+
+    use super::*;
 
     pub fn encode(bytes: impl AsRef<[u8]>) -> Vec<u8> {
         let mut encoder = ZlibEncoder::new(Vec::new(), Compression::fast());
@@ -40,8 +42,9 @@ pub mod deflate {
 }
 
 pub mod brotli {
-    use super::*;
     use ::brotli::{reader::Decompressor as BrotliDecoder, CompressorWriter as BrotliEncoder};
+
+    use super::*;
 
     pub fn encode(bytes: impl AsRef<[u8]>) -> Vec<u8> {
         let mut encoder = BrotliEncoder::new(
@@ -64,8 +67,9 @@ pub mod brotli {
 }
 
 pub mod zstd {
-    use super::*;
     use ::zstd::stream::{read::Decoder, write::Encoder};
+
+    use super::*;
 
     pub fn encode(bytes: impl AsRef<[u8]>) -> Vec<u8> {
         let mut encoder = Encoder::new(Vec::new(), 3).unwrap();