+++ actix-http/src/h1/dispatcher.rs
@@ -1115,6 +1115,7 @@ where
                         let inner = inner.as_mut().project();
                         inner.flags.insert(Flags::READ_DISCONNECT);
                         if let Some(mut payload) = inner.payload.take() {
+                            payload.set_error(io::Error::from(io::ErrorKind::UnexpectedEof).into());
                             payload.feed_eof();
                         }
                     };