+++ actix-http/CHANGES.md
@@ -7,6 +7,10 @@
 - Add `body::to_body_limit()` function.
 - Add `body::BodyLimitExceeded` error type.
 
+### Changed
+
+- Minimum supported Rust version (MSRV) is now 1.65 due to transitive `time` dependency.
+
 ## 3.3.1 - 2023-03-02
 
 ### Fixed