+++ actix-http/CHANGES.md
@@ -7,6 +7,9 @@
 - Add `body::to_body_limit()` function.
 - Add `body::BodyLimitExceeded` error type.
 
+### Fixed
+- Fix truncated body ending without error when connection closed abnormally. [#3067]
+
 ### Changed
 
 - Minimum supported Rust version (MSRV) is now 1.65 due to transitive `time` dependency.