+++ actix-http/CHANGES.md
@@ -9,7 +9,7 @@
 
 ### Changed
 
-- Minimum supported Rust version (MSRV) is now 1.65 due to transitive `time` dependency.
+- Minimum supported Rust version (MSRV) is now 1.68 due to transitive `time` dependency.
 
 ## 3.3.1 - 2023-03-02